+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { ThemeProvider, useTheme } from 'styled-components';
+import { ThemeProvider } from 'styled-components';
 import { Reset } from 'styled-reset';
 import Router from './routes/Router';
 import GlobalStyle from './styles/GlobalStyle';
@@ -10,7 +10,7 @@ import checkTokenValid from './api/tokenValid';
 import loginState from './atoms/login';
 import themeState from './atoms/darkMode';
 import { darkTheme, lightTheme } from './styles/Theme';
-// import { useTheme } from './hooks/useTheme';
+import useTheme from './hooks/useTheme';
 
 function App() {
   const queryClient = new QueryClient();
@@ -18,23 +18,23 @@ function App() {
   const [isLogin, setIsLogin] = useRecoilState(loginState);
   const [themeMode, setThemeMode] = useRecoilState(themeState);
 
-  // // 사용자 OS 설정에 따라 테마 설정
-  // const getInitialTheme = useCallback(() => {
-  //   let theme = localStorage.getItem('theme');
-  //   const invalidTheme = theme !== 'light' && theme !== 'dark';
+  // 사용자 OS 설정에 따라 테마 설정
+  const getInitialTheme = useCallback(() => {
+    let theme = localStorage.getItem('theme');
+    const invalidTheme = theme !== 'light' && theme !== 'dark';
 
-  //   if (!theme || invalidTheme) {
-  //     const { matches } = window.matchMedia('(prefers-color-scheme: dark)');
-  //     theme = matches ? 'dark' : 'light';
-  //   }
-  //   console.log('theme', theme);
+    if (!theme || invalidTheme) {
+      const { matches } = window.matchMedia('(prefers-color-scheme: dark)');
+      theme = matches ? 'dark' : 'light';
+    }
+    console.log('theme', theme);
 
-  //   return theme;
-  // }, []);
+    return theme;
+  }, []);
 
-  // useEffect(() => {
-  //   setThemeMode(getInitialTheme);
-  // }, []);
+  useEffect(() => {
+    setThemeMode(getInitialTheme);
+  }, []);
 
   // console.log(themeMode);
